Control buffer properties when we have a shaping plan
This should improve performance as we're not asking Harfbuzz to work out
information about the text in the buffer when we already have a shaping
plan.

The properties of the buffer are used over those set by the shaping plan,
so to avoid calling UnicodeBuffer::guess_segment_properties we must manually
set those fields. They're currently not accessible from the ShapePlan
directly and it proved to be more ergonomic to pull everything into a
separate (albeit quite redundant-feeling) struct

https://chat.google.com/room/AAAAmakukgk/UhQiK6H9XB0/UhQiK6H9XB0?cls=10
https://github.com/harfbuzz/harfrust/pull/193